% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestImputeSingleData.R
\docType{data}
\name{TestImputeSingleData}
\alias{TestImputeSingleData}
\alias{counts}
\alias{labels}
\title{TestImputeSingleData: A test dataset for ImputeSingle}
\format{\itemize{
  \item counts. A non-negative integer matrix of scRNA-seq raw read counts, rows are genes and columns are cells.
  \item labels. A vector of integer specifying the cell types in the read counts matrix, corresponding to the columns of \code{counts}.
}}
\source{
Petropoulos S, et al. Cell, 2016, 165(4): 1012-1026.
}
\usage{
data(TestImputeSingleData)
}
\description{
A test dataset containing a single-cell RNA-seq (scRNA-seq) read counts matrix and its cell type information.
}
\details{
\itemize{
  \item counts. A matrix of raw read counts of scRNA-seq data which has 200 genes (rows) and 150 cells (columns).
  \item labels. A vector of integer specifying the two cell types in \code{counts}. Also could be generated by: \code{labels <- c(rep(1,50), rep(2,100))}.
}
}
\examples{
# Load test data for ImputeSingle
data(TestImputeSingleData)

# Run ImputeSingle with Kcluster specified
ImputeSingle(counts = counts, Kcluster = 2, parallel = TRUE)

# Run ImputeSingle with labels specified
ImputeSingle(counts = counts, labels = labels, parallel = TRUE)

}
\seealso{
\code{\link{ImputeSingle}}, for the imputation of single-cell RNA-seq data.
}
\keyword{data}
